% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{association_residuals}
\alias{association_residuals}
\title{Extract standardized association residuals}
\usage{
association_residuals(x, min_abs = 0)
}
\arguments{
\item{x}{A \code{biblio_group_comparison}.}

\item{min_abs}{Minimum absolute residual retained.}
}
\value{
Long data frame of group-entity residuals.
}
\description{
Extract standardized association residuals
}
\examples{
z <- compare_groups(as_biblio_project(example_biblio()),rep(c("a","b"),6),permutations=19)
association_residuals(z)
head(association_residuals(z,min_abs=1))
subset(association_residuals(z), residual>0)
}
