% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{audit_biblio}
\alias{audit_biblio}
\title{Inspect the provenance log}
\usage{
audit_biblio(x)
}
\arguments{
\item{x}{A \code{biblio_project}.}
}
\value{
A data frame.
}
\description{
Inspect the provenance log
}
\examples{
audit_biblio(as_biblio_project(example_biblio()))
nrow(audit_biblio(as_biblio_project(head(example_biblio()))))
names(audit_biblio(as_biblio_project(example_biblio())))
}
