% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks_storage.R
\name{backend_status}
\alias{backend_status}
\title{Optional backend status}
\usage{
backend_status()
}
\value{
A data frame indicating installed scalable backends.
}
\description{
Optional backend status
}
\examples{
backend_status()
subset(backend_status(), available)
backend_status()$backend
}
