% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_ui.R
\name{biblio_app}
\alias{biblio_app}
\title{Interactive bibliometric workbench}
\usage{
biblio_app(data = NULL)
}
\arguments{
\item{data}{Optional initial data frame or \code{biblio_project}.}
}
\value{
A Shiny application object.
}
\description{
Interactive bibliometric workbench
}
\examples{
\dontrun{
biblio_app()
biblio_app(example_biblio())
shiny::runApp(biblio_app(as_biblio_project(example_biblio())))
}
}
