% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{biblio_health}
\alias{biblio_health}
\title{Diagnose corpus quality}
\usage{
biblio_health(x)
}
\arguments{
\item{x}{A \code{biblio_project}.}
}
\value{
A data frame of checks and counts.
}
\description{
Diagnose corpus quality
}
\examples{
biblio_health(as_biblio_project(example_biblio()))
biblio_health(as_biblio_project(head(example_biblio())))
subset(biblio_health(as_biblio_project(example_biblio())), n>0)
}
