% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{biblio_metrics}
\alias{biblio_metrics}
\title{Author-level bibliometric metrics}
\usage{
biblio_metrics(x)
}
\arguments{
\item{x}{A \code{biblio_project}.}
}
\value{
Author publication, citation, h-index, g-index and m-index metrics.
}
\description{
Author-level bibliometric metrics
}
\examples{
biblio_metrics(as_biblio_project(example_biblio()))
head(biblio_metrics(as_biblio_project(example_biblio())),3)
subset(biblio_metrics(as_biblio_project(example_biblio())), documents>=2)
}
