% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{biblio_plot_annual}
\alias{biblio_plot_annual}
\title{Annual production barplot (base graphics, no extra packages)}
\usage{
biblio_plot_annual(x)
}
\arguments{
\item{x}{A \code{biblio_project}.}
}
\value{
The plotted data, invisibly.
}
\description{
Annual production barplot (base graphics, no extra packages)
}
\examples{
biblio_plot_annual(as_biblio_project(example_biblio()))
}
