% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{biblio_plot_citations}
\alias{biblio_plot_citations}
\title{Citation distribution histogram}
\usage{
biblio_plot_citations(x)
}
\arguments{
\item{x}{A \code{biblio_project}.}
}
\value{
The histogram object, invisibly.
}
\description{
Citation distribution histogram
}
\examples{
biblio_plot_citations(as_biblio_project(example_biblio()))
}
