% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{biblio_plot_sources}
\alias{biblio_plot_sources}
\title{Top sources barplot}
\usage{
biblio_plot_sources(x, top_n = 15L)
}
\arguments{
\item{x}{A \code{biblio_project}.}

\item{top_n}{Number of sources.}
}
\value{
The plotted data, invisibly.
}
\description{
Top sources barplot
}
\examples{
biblio_plot_sources(as_biblio_project(example_biblio()))
}
