% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{biblio_plot_terms}
\alias{biblio_plot_terms}
\title{Top terms barplot}
\usage{
biblio_plot_terms(x, top_n = 15L, field = c("title", "abstract"))
}
\arguments{
\item{x}{A \code{biblio_project}.}

\item{top_n}{Number of terms.}

\item{field}{\code{"title"} or \code{"abstract"}.}
}
\value{
The plotted data, invisibly.
}
\description{
Top terms barplot
}
\examples{
biblio_plot_terms(as_biblio_project(example_biblio()))
}
