% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biblioIntegrator-package.R
\docType{package}
\name{biblioIntegrator-package}
\alias{biblioIntegrator-package}
\alias{biblioIntegrator}
\title{biblioIntegrator: Harmonized Comparative Bibliometrics}
\description{
An R-first framework for bibliographic harmonization, provenance, descriptive
bibliometrics, comparative inference, network analysis, temporal/text analysis,
scalable storage and optional Python integration.
}
\author{
\strong{Maintainer}: Walter Esfrain Pereira \email{walterufpb@yahoo.com.br} (\href{https://orcid.org/0000-0003-1085-0191}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Walter Esfrain Pereira \email{walterufpb@yahoo.com.br} (\href{https://orcid.org/0000-0003-1085-0191}{ORCID}) [copyright holder]
  \item Magali Haidee Pereira Martinez (\href{https://orcid.org/0009-0009-5419-959X}{ORCID})
}

}
\keyword{internal}
