% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_text.R
\name{citation_trajectory}
\alias{citation_trajectory}
\title{Citation trajectory}
\usage{
citation_trajectory(x, current_year = as.integer(format(Sys.Date(), "\%Y")))
}
\arguments{
\item{x}{A \code{biblio_project}.}

\item{current_year}{Reference year.}
}
\value{
Work-level age, citations and velocity.
}
\description{
Citation trajectory
}
\examples{
citation_trajectory(as_biblio_project(example_biblio()))
head(citation_trajectory(as_biblio_project(example_biblio())),3)
citation_trajectory(as_biblio_project(example_biblio()),current_year=2026)
}
