% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{citation_velocity}
\alias{citation_velocity}
\title{Citation velocity}
\usage{
citation_velocity(x, current_year = as.integer(format(Sys.Date(), "\%Y")))
}
\arguments{
\item{x}{A \code{biblio_project}.}

\item{current_year}{Reference year.}
}
\value{
Work-level citations per year.
}
\description{
Citation velocity
}
\examples{
citation_velocity(as_biblio_project(example_biblio()))
head(citation_velocity(as_biblio_project(example_biblio())),3)
summary(citation_velocity(as_biblio_project(example_biblio()))$velocity)
}
