% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{compare_sources}
\alias{compare_sources}
\title{Compare coverage across bibliographic sources}
\usage{
compare_sources(...)
}
\arguments{
\item{...}{Data frames or \code{biblio_project} objects.}
}
\value{
A list with coverage and pairwise overlap.
}
\description{
Compare coverage across bibliographic sources
}
\examples{
compare_sources(example_biblio(), head(example_biblio(),8))
compare_sources(as_biblio_project(example_biblio()), example_biblio()[5:12,])
compare_sources(A=example_biblio(), B=example_biblio()[1:6,])
}
