% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{describe_biblio}
\alias{describe_biblio}
\title{Descriptive bibliometric summary}
\usage{
describe_biblio(x)
}
\arguments{
\item{x}{A \code{biblio_project}.}
}
\value{
A list of summary tables.
}
\description{
Descriptive bibliometric summary
}
\examples{
describe_biblio(as_biblio_project(example_biblio()))
describe_biblio(as_biblio_project(head(example_biblio())))$annual
describe_biblio(as_biblio_project(example_biblio()))$top_sources
}
