% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{enable_python_backend}
\alias{enable_python_backend}
\title{Enable an existing Python backend}
\usage{
enable_python_backend(python)
}
\arguments{
\item{python}{Python executable.}
}
\value{
Backend status.
}
\description{
Enable an existing Python backend
}
\examples{
\dontrun{
enable_python_backend("path/to/python")
enable_python_backend(Sys.which("python"))
enable_python_backend(reticulate::virtualenv_python("r-bibliointegrator"))
}
}
