% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_ui.R
\name{export_vosviewer}
\alias{export_vosviewer}
\title{Export a network for VOSviewer}
\usage{
export_vosviewer(graph, path)
}
\arguments{
\item{graph}{An igraph network.}

\item{path}{Output network text file.}
}
\value{
Output path invisibly.
}
\description{
Export a network for VOSviewer
}
\examples{
x <- as_biblio_project(example_biblio())
g <- bibliographic_network(x, "coauthor")
f <- tempfile()
export_vosviewer(g, f)
file.exists(f)
g <- bibliographic_network(x, "keyword")
f <- tempfile()
export_vosviewer(g, f)
readLines(f, n = 2)
g <- bibliographic_network(x, "coauthor")
f <- tempfile()
invisible(export_vosviewer(g, f))
}
