% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis.R
\name{fetch_opencitations}
\alias{fetch_opencitations}
\title{Fetch OpenCitations Index relations}
\usage{
fetch_opencitations(identifier, direction = c("citations", "references"))
}
\arguments{
\item{identifier}{DOI or supported persistent identifier.}

\item{direction}{\code{"citations"} or \code{"references"}.}
}
\value{
A data frame returned by the OpenCitations API.
}
\description{
Fetch OpenCitations Index relations
}
\examples{
\dontrun{
fetch_opencitations("10.1038/nature12373")
fetch_opencitations("10.1038/nature12373",direction="references")
head(fetch_opencitations("10.1038/nature12373"))
}
}
