% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{group_ca}
\alias{group_ca}
\title{Correspondence analysis of group-entity associations}
\usage{
group_ca(x, ndim = 2)
}
\arguments{
\item{x}{A \code{biblio_group_comparison}.}

\item{ndim}{Number of dimensions.}
}
\value{
Row/column coordinates and singular values.
}
\description{
Correspondence analysis of group-entity associations
}
\examples{
z <- compare_groups(as_biblio_project(example_biblio()),rep(c("a","b"),6),permutations=19)
group_ca(z)
group_ca(z,ndim=1)$rows
group_ca(z)$singular_values
}
