% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{group_mca}
\alias{group_mca}
\title{Multiple correspondence analysis of group and entity presence}
\usage{
group_mca(x, groups, entity = c("keyword", "author"), ncp = 2)
}
\arguments{
\item{x}{A \code{biblio_project}.}

\item{groups}{Group definition.}

\item{entity}{Entity type.}

\item{ncp}{Number of dimensions.}
}
\value{
A \code{FactoMineR::MCA} result.
}
\description{
Multiple correspondence analysis of group and entity presence
}
\examples{
\donttest{
x <- as_biblio_project(example_biblio()); g <- rep(c("a","b"),6)
if (requireNamespace("FactoMineR",quietly=TRUE)) group_mca(x,g,ncp=2)
if (requireNamespace("FactoMineR",quietly=TRUE)) group_mca(x,cbind(a=1:12<=7,b=1:12>=5),ncp=1)
if (requireNamespace("FactoMineR",quietly=TRUE)) names(group_mca(x,g,ncp=2))
}
}
