% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks_storage.R
\name{network_centrality}
\alias{network_centrality}
\title{Network centrality table}
\usage{
network_centrality(graph)
}
\arguments{
\item{graph}{An igraph object.}
}
\value{
A data frame of degree, strength, betweenness and PageRank.
}
\description{
Network centrality table
}
\examples{
g <- bibliographic_network(as_biblio_project(example_biblio()),"coauthor"); network_centrality(g)
head(network_centrality(g),3)
network_centrality(bibliographic_network(as_biblio_project(example_biblio()),"keyword"))
}
