% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks_storage.R
\name{network_communities}
\alias{network_communities}
\title{Detect network communities}
\usage{
network_communities(graph, method = c("louvain", "walktrap", "label_prop"))
}
\arguments{
\item{graph}{Undirected igraph network.}

\item{method}{\code{"louvain"}, \code{"walktrap"}, or \code{"label_prop"}.}
}
\value{
A data frame with node membership.
}
\description{
Detect network communities
}
\examples{
g <- bibliographic_network(as_biblio_project(example_biblio()),"coauthor"); network_communities(g)
network_communities(g,"walktrap")
network_communities(g,"label_prop")
}
