% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{normalized_citations}
\alias{normalized_citations}
\title{Field/year normalized citations}
\usage{
normalized_citations(x, strata = "year")
}
\arguments{
\item{x}{A \code{biblio_project}.}

\item{strata}{Character columns in \code{works} defining comparable strata.}
}
\value{
Work-level normalized citation scores.
}
\description{
Field/year normalized citations
}
\examples{
normalized_citations(as_biblio_project(example_biblio()))
head(normalized_citations(as_biblio_project(example_biblio())),4)
mean(normalized_citations(as_biblio_project(example_biblio()))$normalized,na.rm=TRUE)
}
