% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{print.biblio_group_comparison}
\alias{print.biblio_group_comparison}
\title{Print a group comparison}
\usage{
\method{print}{biblio_group_comparison}(x, ...)
}
\arguments{
\item{x}{A \code{biblio_group_comparison}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a group comparison
}
