% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{print.biblio_project}
\alias{print.biblio_project}
\title{Print a bibliometric project}
\usage{
\method{print}{biblio_project}(x, ...)
}
\arguments{
\item{x}{A \code{biblio_project}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a bibliometric project
}
