% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_text.R
\name{temporal_growth}
\alias{temporal_growth}
\title{Annual bibliometric growth}
\usage{
temporal_growth(x)
}
\arguments{
\item{x}{A \code{biblio_project}.}
}
\value{
Annual documents, citations and year-over-year document growth.
}
\description{
Annual bibliometric growth
}
\examples{
temporal_growth(as_biblio_project(example_biblio()))
tail(temporal_growth(as_biblio_project(example_biblio())),3)
summary(temporal_growth(as_biblio_project(example_biblio()))$documents)
}
