% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{to_bibliometrix}
\alias{to_bibliometrix}
\title{Convert to bibliometrix field-tag data}
\usage{
to_bibliometrix(x)
}
\arguments{
\item{x}{A \code{biblio_project}.}
}
\value{
A data frame compatible with common \code{bibliometrix} workflows.
}
\description{
Convert to bibliometrix field-tag data
}
\examples{
to_bibliometrix(as_biblio_project(example_biblio()))
nrow(to_bibliometrix(as_biblio_project(head(example_biblio()))))
names(to_bibliometrix(as_biblio_project(example_biblio())))
}
