% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{to_biblium}
\alias{to_biblium}
\title{Create a Biblium-ready data frame}
\usage{
to_biblium(x)
}
\arguments{
\item{x}{A \code{biblio_project}.}
}
\value{
A data frame with Biblium canonical field names.
}
\description{
Create a Biblium-ready data frame
}
\examples{
to_biblium(as_biblio_project(example_biblio()))
head(to_biblium(as_biblio_project(example_biblio())),3)
names(to_biblium(as_biblio_project(example_biblio())))
}
