% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_text.R
\name{trend_topics}
\alias{trend_topics}
\title{Topic trajectories by year}
\usage{
trend_topics(x, min_total = 1)
}
\arguments{
\item{x}{A \code{biblio_project}.}

\item{min_total}{Minimum corpus-wide keyword frequency.}
}
\value{
Long data frame of keyword counts by year.
}
\description{
Topic trajectories by year
}
\examples{
trend_topics(as_biblio_project(example_biblio()))
trend_topics(as_biblio_project(example_biblio()),min_total=2)
head(trend_topics(as_biblio_project(example_biblio())),4)
}
