% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{deduplicate_biblio}
\alias{deduplicate_biblio}
\title{Deduplicate bibliographic records}
\usage{
deduplicate_biblio(x, method = c("doi_title_year"))
}
\arguments{
\item{x}{A \code{biblio_project}.}

\item{method}{Matching hierarchy: DOI then normalized title-year.}
}
\value{
A deduplicated \code{biblio_project} with a log attribute.
}
\description{
Deduplicate bibliographic records
}
\examples{
x <- as_biblio_project(rbind(example_biblio(),example_biblio()[1,])); deduplicate_biblio(x)
nrow(deduplicate_biblio(x)$works)
attr(deduplicate_biblio(x),"dedup_log")
}
