% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{install_biblium_backend}
\alias{install_biblium_backend}
\title{Install an isolated Biblium backend}
\usage{
install_biblium_backend(
  envname = "r-bibliointegrator",
  python = NULL,
  version = "2.16.0"
)
}
\arguments{
\item{envname}{Virtual environment name or path.}

\item{python}{Python executable used to create the environment.}

\item{version}{Biblium version.}
}
\value{
Invisible virtual environment path.
}
\description{
Install an isolated Biblium backend
}
\examples{
\dontrun{
install_biblium_backend()
install_biblium_backend(version="2.16.0")
install_biblium_backend(envname="r-bibliointegrator")
}
}
