% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_text.R
\name{term_frequency}
\alias{term_frequency}
\title{Term frequency from titles or abstracts}
\usage{
term_frequency(
  x,
  field = c("title", "abstract"),
  stopwords = c("and", "the", "for", "with", "under", "of", "in", "to")
)
}
\arguments{
\item{x}{A \code{biblio_project}.}

\item{field}{\code{"title"} or \code{"abstract"}.}

\item{stopwords}{Optional vector removed before counting.}
}
\value{
Term-frequency data frame.
}
\description{
Term frequency from titles or abstracts
}
\examples{
term_frequency(as_biblio_project(example_biblio()))
head(term_frequency(as_biblio_project(example_biblio()),stopwords=c("and","the")),5)
term_frequency(as_biblio_project(example_biblio()),field="abstract")
}
